Clean useless code
