+++ app/src/main/java/com/example/doggo_android/views/SettingsFragment.java
@@ -1,0 +1,30 @@
+package com.example.doggo_android.views;
+
+import android.os.Bundle;
+
+import androidx.fragment.app.Fragment;
+
+import android.view.LayoutInflater;
+import android.view.View;
+import android.view.ViewGroup;
+
+import com.example.doggo_android.R;
+
+
+public class SettingsFragment extends Fragment {
+
+
+    public SettingsFragment() {
+    }
+
+    @Override
+    public void onCreate(Bundle savedInstanceState) {
+        super.onCreate(savedInstanceState);
+    }
+
+    @Override
+    public View onCreateView(LayoutInflater inflater, ViewGroup container,
+                             Bundle savedInstanceState) {
+        return inflater.inflate(R.layout.fragment_settings, container, false);
+    }
+}
