+++ app/src/main/res/layout/fragment_contact.xml
@@ -3,7 +3,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".ContactFragment"
+    tools:context=".views.ContactFragment"
     android:background="@color/red">
 
     <!-- TODO: Update blank fragment layout -->
