+++ app/src/main/res/layout/fragment_home.xml
@@ -4,7 +4,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:background="@color/teal_200"
-    tools:context=".HomeFragment">
+    tools:context=".views.HomeFragment">
 
     <!-- TODO: Update blank fragment layout -->
     <TextView
