+++ app/src/main/res/layout/fragment_members.xml
@@ -3,7 +3,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".MembersFragment"
+    tools:context=".views.MembersFragment"
     android:background="@color/teal_700">
 
     <!-- TODO: Update blank fragment layout -->
