+++ app/src/main/res/layout/fragment_profile.xml
@@ -4,7 +4,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:background="@color/purple_200"
-    tools:context=".ProfileFragment">
+    tools:context=".views.ProfileFragment">
 
     <!-- TODO: Update blank fragment layout -->
     <TextView
