+++ app/src/main/res/layout/fragment_settings.xml
@@ -4,7 +4,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:background="@color/purple_700"
-    tools:context=".SettingsFragment">
+    tools:context=".views.SettingsFragment">
 
     <!-- TODO: Update blank fragment layout -->
     <TextView
